requirements.txt
streamlit
jinja2
python-docx
nltk
transformers
accelerate
bitsandbytes
torch
sentencepiece
